requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.4
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
freeze==3.0
idna==2.10
isodate==0.6.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
